+++ src/app/u/components/home/profile/ProfileSettings/ProfileSettings.tsx
@@ -1,7 +1,5 @@
 "use client";
 
-import UnderlineInput from '@/app/components/UnderlineInput';
-import UnderlineSelect from '@/app/components/UnderlineSelect';
 import { updateUserPreferences } from '@/services/firebase/update';
 import CurrencySelector from './ProfileCurrencySelector';
