+++ src/utils/generate-random.ts
@@ -1,0 +1,9 @@
+const chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+
+export const generateRandomChars = (numChar: number) => {
+    let code = '';
+    for (let i = 0; i < numChar; i++) {
+        code += chars.charAt(Math.floor(Math.random() * chars.length));
+    }
+    return code;
+};
